Teleport to pointer movement and hand collisions
Added teleporting to wherever the player wants to teleport. For the quest, you push to right thumbstick forward to highlight where you want to teleport and release to teleport to that area. Also added hand collisions and a timer that would enable collisions again if the player lets go of the item after a certain period. Also moved some script folders for organization.