Inventory System Update
Added the UI for the inventory system. Can be brought up by pressing the A button or the X button on the Quest Controller. Inventory shows up as a list of items where clicking on them spawns the item in front of the player. Adding items to inventory could be done by grabbing the item and moving it behind your head. When shooting with the gun, UI now displays how many seeds the player has in their inventory (could be revamped later). When reloading, seed ammo is removed from inventory and added to gun based on the remaining or max ammo.