Integrate Creature AI System Task
Added a state machine that manages states between the enemy AI being aggressive and passive. Currently when passive, the AI returns to their spawn point instead of wandering. By default, the AI is set to aggressive and would attack the player. When the player "kills" the AI, the AI is set to passive. You could test it out in the lab scene.