Separated VR Grab functionality from gun
Separated the VR grab functionality from the shooting and setup functionality. There's an issue currently where the gun doesn't accurately follow the player 1:1, but I compared this with the develop branch before I made the changes and it seemed like it was there before I worked on this change. Currently, the VR grab functionality has been disabled until we get back to the VR version of the game at some point.