User able to holster gun (Primary object) and swap between two
The player guns in the lab scene are able to be picked up and held onto by the player, currently their are a couple issues
when both guns are picked up and swapped, they will no longer fire.
there seems to be an issue with unequipping the second gun after swapping.